[refactor] Custom Exception Package 변경
