+++ src/main/java/kr/startoff/backend/exception/custom/InvalidPasswordException.java
@@ -1,4 +1,4 @@
-package kr.startoff.backend.exception;
+package kr.startoff.backend.exception.custom;
 
 public class InvalidPasswordException extends RuntimeException {
 	public InvalidPasswordException() {
