+++ src/main/java/kr/startoff/backend/exception/custom/OAuth2AuthenticationProcessingException.java
@@ -1,4 +1,4 @@
-package kr.startoff.backend.exception;
+package kr.startoff.backend.exception.custom;
 
 public class OAuth2AuthenticationProcessingException extends RuntimeException {
 	public OAuth2AuthenticationProcessingException(String s) {
