+++ src/main/java/kr/startoff/backend/service/UserDetailsServiceImpl.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import kr.startoff.backend.entity.User;
-import kr.startoff.backend.exception.UserNotFoundException;
+import kr.startoff.backend.exception.custom.UserNotFoundException;
 import kr.startoff.backend.repository.UserRepository;
 import kr.startoff.backend.security.UserPrincipal;
 import lombok.RequiredArgsConstructor;
